requirements.txt
Flask==2.3.3
boto3==1.34.76
mysql-connector-python==8.3.0
APScheduler==3.10.4
requests==2.31.0
